+++ Config/ConfigLoader.php
@@ -25,6 +25,7 @@ namespace Bartacus\Bundle\BartacusBundle\Config;
 
 use Bartacus\Bundle\BartacusBundle\ContentElement\Loader\ContentElementConfigLoader;
 use Bartacus\Bundle\BartacusBundle\Middleware\PrepareContentElementRenderer;
+use Bartacus\Bundle\BartacusBundle\Middleware\SymfonyRouteResolver;
 
 /**
  * Delegating central config loader called on various places within TYPO3
@@ -51,6 +52,21 @@ class ConfigLoader
     {
         return [
             'frontend' => [
+                'bartacus/symfony-route-resolver' => [
+                    'target' => SymfonyRouteResolver::class,
+                    'after' => [
+                        'typo3/cms-frontend/authentication',
+                        'typo3/cms-frontend/backend-user-authentication',
+                        'typo3/cms-frontend/tsfe',
+                        'typo3/cms-frontend/site',
+                        'typo3/cms-frontend/base-redirect-resolver',
+                        'typo3/cms-frontend/static-route-resolver',
+                        'typo3/cms-redirects/redirecthandler',
+                    ],
+                    'before' => [
+                        'typo3/cms-frontend/page-resolver',
+                    ],
+                ],
                 'bartacus/prepare-content-element-renderer' => [
                     'target' => PrepareContentElementRenderer::class,
                     'after' => [