+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 - Handle TYPO3 page rendering with all the usual Symfony events around as master request
 - Dispatch all the usual Symfony events around each content element as sub request
+- Handle Symfony routing before the page resolver
 
 ## [2.0.1] - 2019-04-12
 ### Added